Refactor(performance): Split the batch module
batch.py held 719 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the priority
and retry enums plus the request/result records), rate_limit (the
RateLimitOptimizer budget tracking), batcher (RequestQueue and
RequestBatcher, deciding what runs next) and processor (BatchProcessor
wiring the three together, with its module-level wrappers). The
package __init__ re-exports the previous names.
